+++ web_src/src/pages/factories/pages/missions/MissionsWorkOrdersPage.tsx
@@ -49,9 +49,9 @@ export function MissionsWorkOrdersPage() {
   const handleAssigneesSave = async (orderId: string, assigneeIds: string[]) => {
     try {
       await updateAssignees.mutateAsync({ orderId, assigneeIds });
-      showSuccessToast("Assignees updated.");
+      showSuccessToast("Owners updated.");
     } catch {
-      showErrorToast("Failed to update assignees.");
+      showErrorToast("Failed to update owners.");
     }
   };
